third_party/blink/web_tests/platform/generic/fast/events/touch/touch-slider-expected.txt

Tests that the touch events originating on an input element with type=range update the slider position. This test is only expected to pass if ENABLE_TOUCH_SLIDER is defined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.touches[0].target.id is "slider1"
PASS sliderValue is "50"
PASS sliderValue is "50"
PASS sliderValue is "0"
PASS sliderValue is "100"
PASS sliderValue is "50"
PASS event.touches[0].target.id is "slider2"
PASS sliderValue is "50"
PASS sliderValue is "50"
PASS sliderValue is "0"
PASS sliderValue is "100"
PASS sliderValue is "50"
PASS event.touches[0].target.id is "slider3"
PASS sliderValue is "50"
PASS sliderValue is "50"
PASS sliderValue is "0"
PASS sliderValue is "100"
PASS sliderValue is "50"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
